abstracted cors config methods into their own class
